requirements.txt
panda3d==1.10.8
DirectFolderBrowser
DirectGuiExtension
